[~] Move Panel to molecules
