+++ src/components/atoms/Icons/Expand/Expand.react.js
