+++ src/components/atoms/Icons/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
+import IconExpand from './Expand/Expand.react';
 import Icon from './Icon.react';
 
-const compress = `M21.71 2.29C21.617 2.19627 21.5064 2.12188 21.3846 2.07111C21.2627 2.02034 21.132 1.9942 21 1.9942C20.868 1.9942 20.7373 2.02034 20.6154 2.07111C20.4936 2.12188 20.383 2.19627 20.29 2.29L14.5 8.09V6.5C14.5 6.23478 14.3946 5.98043 14.2071 5.79289C14.0196 5.60536 13.7652 5.5 13.5 5.5C13.2348 5.5 12.9804 5.60536 12.7929 5.79289C12.6054 5.98043 12.5 6.23478 12.5 6.5V10.5C12.5016 10.6307 12.5288 10.7598 12.58 10.88C12.6815 11.1243 12.8757 11.3185 13.12 11.42C13.2402 11.4712 13.3693 11.4984 13.5 11.5H17.5C17.7652 11.5 18.0196 11.3946 18.2071 11.2071C18.3946 11.0196 18.5 10.7652 18.5 10.5C18.5 10.2348 18.3946 9.98043 18.2071 9.79289C18.0196 9.60536 17.7652 9.5 17.5 9.5H15.91L21.71 3.71C21.8037 3.61704 21.8781 3.50644 21.9289 3.38458C21.9797 3.26272 22.0058 3.13201 22.0058 3C22.0058 2.86799 21.9797 2.73728 21.9289 2.61542C21.8781 2.49356 21.8037 2.38296 21.71 2.29V2.29ZM10.88 12.58C10.7598 12.5288 10.6307 12.5016 10.5 12.5H6.5C6.23478 12.5 5.98043 12.6054 5.79289 12.7929C5.60536 12.9804 5.5 13.2348 5.5 13.5C5.5 13.7652 5.60536 14.0196 5.79289 14.2071C5.98043 14.3946 6.23478 14.5 6.5 14.5H8.09L2.29 20.29C2.19627 20.383 2.12188 20.4936 2.07111 20.6154C2.02034 20.7373 1.9942 20.868 1.9942 21C1.9942 21.132 2.02034 21.2627 2.07111 21.3846C2.12188 21.5064 2.19627 21.617 2.29 21.71C2.38296 21.8037 2.49356 21.8781 2.61542 21.9289C2.73728 21.9797 2.86799 22.0058 3 22.0058C3.13201 22.0058 3.26272 21.9797 3.38458 21.9289C3.50644 21.8781 3.61704 21.8037 3.71 21.71L9.5 15.91V17.5C9.5 17.7652 9.60536 18.0196 9.79289 18.2071C9.98043 18.3946 10.2348 18.5 10.5 18.5C10.7652 18.5 11.0196 18.3946 11.2071 18.2071C11.3946 18.0196 11.5 17.7652 11.5 17.5V13.5C11.4984 13.3693 11.4712 13.2402 11.42 13.12C11.3185 12.8757 11.1243 12.6815 10.88 12.58Z`;
-const expand = `M9.79 12.79L4 18.59V17C4 16.7348 3.89464 16.4804 3.70711 16.2929C3.51957 16.1054 3.26522 16 3 16C2.73478 16 2.48043 16.1054 2.29289 16.2929C2.10536 16.4804 2 16.7348 2 17V21C2.00158 21.1307 2.02876 21.2598 2.08 21.38C2.18147 21.6243 2.37565 21.8185 2.62 21.92C2.74022 21.9712 2.86932 21.9984 3 22H7C7.26522 22 7.51957 21.8946 7.70711 21.7071C7.89464 21.5196 8 21.2652 8 21C8 20.7348 7.89464 20.4804 7.70711 20.2929C7.51957 20.1054 7.26522 20 7 20H5.41L11.21 14.21C11.3983 14.0217 11.5041 13.7663 11.5041 13.5C11.5041 13.2337 11.3983 12.9783 11.21 12.79C11.0217 12.6017 10.7663 12.4959 10.5 12.4959C10.2337 12.4959 9.9783 12.6017 9.79 12.79V12.79ZM21.92 2.62C21.8185 2.37565 21.6243 2.18147 21.38 2.08C21.2598 2.02876 21.1307 2.00158 21 2H17C16.7348 2 16.4804 2.10536 16.2929 2.29289C16.1054 2.48043 16 2.73478 16 3C16 3.26522 16.1054 3.51957 16.2929 3.70711C16.4804 3.89464 16.7348 4 17 4H18.59L12.79 9.79C12.6963 9.88296 12.6219 9.99356 12.5711 10.1154C12.5203 10.2373 12.4942 10.368 12.4942 10.5C12.4942 10.632 12.5203 10.7627 12.5711 10.8846C12.6219 11.0064 12.6963 11.117 12.79 11.21C12.883 11.3037 12.9936 11.3781 13.1154 11.4289C13.2373 11.4797 13.368 11.5058 13.5 11.5058C13.632 11.5058 13.7627 11.4797 13.8846 11.4289C14.0064 11.3781 14.117 11.3037 14.21 11.21L20 5.41V7C20 7.26522 20.1054 7.51957 20.2929 7.70711C20.4804 7.89464 20.7348 8 21 8C21.2652 8 21.5196 7.89464 21.7071 7.70711C21.8946 7.51957 22 7.26522 22 7V3C21.9984 2.86932 21.9712 2.74022 21.92 2.62Z`;
 const filter = `M19 2H5C4.20435 2 3.44129 2.31607 2.87868 2.87868C2.31607 3.44129 2 4.20435 2 5V6.17C1.99986 6.58294 2.08497 6.99147 2.25 7.37V7.43C2.39128 7.75097 2.59139 8.04266 2.84 8.29L9 14.41V21C8.99966 21.1699 9.04264 21.3372 9.12487 21.4859C9.20711 21.6346 9.32589 21.7599 9.47 21.85C9.62914 21.9486 9.81277 22.0006 10 22C10.1565 21.9991 10.3107 21.9614 10.45 21.89L14.45 19.89C14.6149 19.8069 14.7536 19.6798 14.8507 19.5227C14.9478 19.3656 14.9994 19.1847 15 19V14.41L21.12 8.29C21.3686 8.04266 21.5687 7.75097 21.71 7.43V7.37C21.8888 6.99443 21.9876 6.58578 22 6.17V5C22 4.20435 21.6839 3.44129 21.1213 2.87868C20.5587 2.31607 19.7956 2 19 2ZM13.29 13.29C13.1973 13.3834 13.124 13.4943 13.0742 13.6161C13.0245 13.7379 12.9992 13.8684 13 14V18.38L11 19.38V14C11.0008 13.8684 10.9755 13.7379 10.9258 13.6161C10.876 13.4943 10.8027 13.3834 10.71 13.29L5.41 8H18.59L13.29 13.29ZM20 6H4V5C4 4.73478 4.10536 4.48043 4.29289 4.29289C4.48043 4.10536 4.73478 4 5 4H19C19.2652 4 19.5196 4.10536 19.7071 4.29289C19.8946 4.48043 20 4.73478 20 5V6Z`;
 const moon = `M21.64 13C21.4956 12.8807 21.3207 12.8043 21.1351 12.7796C20.9494 12.7548 20.7606 12.7827 20.59 12.86C19.5326 13.3439 18.3829 13.5929 17.22 13.59C15.0689 13.5874 13.006 12.7344 11.4811 11.2171C9.9563 9.69976 9.09319 7.64111 9.08 5.49C9.08456 4.81583 9.16847 4.14455 9.33 3.49C9.36428 3.31554 9.35142 3.13509 9.29275 2.96725C9.23409 2.79941 9.13174 2.65025 8.99624 2.53513C8.86074 2.42001 8.69701 2.3431 8.5219 2.31231C8.34678 2.28153 8.16663 2.29799 8 2.36C6.43232 3.06491 5.06957 4.15676 4.03976 5.53298C3.00994 6.90921 2.34688 8.52463 2.11285 10.2275C1.87882 11.9303 2.0815 13.6648 2.70187 15.2678C3.32223 16.8708 4.33991 18.2898 5.6593 19.3915C6.97869 20.4932 8.55647 21.2414 10.2444 21.5658C11.9324 21.8903 13.6752 21.7803 15.3089 21.2462C16.9427 20.7122 18.414 19.7716 19.5844 18.5128C20.7548 17.254 21.5861 15.7183 22 14.05C22.0504 13.859 22.0431 13.6573 21.9791 13.4705C21.915 13.2836 21.797 13.1199 21.64 13ZM12.14 19.69C10.4618 19.6782 8.8282 19.1479 7.46299 18.1717C6.09778 17.1956 5.06766 15.8213 4.51368 14.2371C3.9597 12.6529 3.90893 10.9362 4.36834 9.322C4.82774 7.70782 5.77485 6.2751 7.08 5.22V5.49C7.08265 8.17848 8.15181 10.7561 10.0529 12.6571C11.9539 14.5582 14.5315 15.6274 17.22 15.63C17.9259 15.6326 18.63 15.5588 19.32 15.41C18.6299 16.7155 17.5965 17.8078 16.3312 18.5691C15.0658 19.3304 13.6167 19.7318 12.14 19.73V19.69Z`;
 const pause = `M16 2C15.2044 2 14.4413 2.31607 13.8787 2.87868C13.3161 3.44129 13 4.20435 13 5V19C13 19.7956 13.3161 20.5587 13.8787 21.1213C14.4413 21.6839 15.2044 22 16 22C16.7956 22 17.5587 21.6839 18.1213 21.1213C18.6839 20.5587 19 19.7956 19 19V5C19 4.20435 18.6839 3.44129 18.1213 2.87868C17.5587 2.31607 16.7956 2 16 2ZM17 19C17 19.2652 16.8946 19.5196 16.7071 19.7071C16.5196 19.8946 16.2652 20 16 20C15.7348 20 15.4804 19.8946 15.2929 19.7071C15.1054 19.5196 15 19.2652 15 19V5C15 4.73478 15.1054 4.48043 15.2929 4.29289C15.4804 4.10536 15.7348 4 16 4C16.2652 4 16.5196 4.10536 16.7071 4.29289C16.8946 4.48043 17 4.73478 17 5V19ZM8 2C7.20435 2 6.44129 2.31607 5.87868 2.87868C5.31607 3.44129 5 4.20435 5 5V19C5 19.7956 5.31607 20.5587 5.87868 21.1213C6.44129 21.6839 7.20435 22 8 22C8.79565 22 9.55871 21.6839 10.1213 21.1213C10.6839 20.5587 11 19.7956 11 19V5C11 4.20435 10.6839 3.44129 10.1213 2.87868C9.55871 2.31607 8.79565 2 8 2ZM9 19C9 19.2652 8.89464 19.5196 8.70711 19.7071C8.51957 19.8946 8.26522 20 8 20C7.73478 20 7.48043 19.8946 7.29289 19.7071C7.10536 19.5196 7 19.2652 7 19V5C7 4.73478 7.10536 4.48043 7.29289 4.29289C7.48043 4.10536 7.73478 4 8 4C8.26522 4 8.51957 4.10536 8.70711 4.29289C8.89464 4.48043 9 4.73478 9 5V19Z`;
@@ -14,8 +13,6 @@ const signal = `M6 15C5.73478 15 5.48043 15.1054 5.29289 15.2929C5.10536 15.4804
 const stop = `M17 2H7C5.67392 2 4.40215 2.52678 3.46447 3.46447C2.52678 4.40215 2 5.67392 2 7V17C2 18.3261 2.52678 19.5979 3.46447 20.5355C4.40215 21.4732 5.67392 22 7 22H17C18.3261 22 19.5979 21.4732 20.5355 20.5355C21.4732 19.5979 22 18.3261 22 17V7C22 5.67392 21.4732 4.40215 20.5355 3.46447C19.5979 2.52678 18.3261 2 17 2ZM20 17C20 17.7956 19.6839 18.5587 19.1213 19.1213C18.5587 19.6839 17.7956 20 17 20H7C6.20435 20 5.44129 19.6839 4.87868 19.1213C4.31607 18.5587 4 17.7956 4 17V7C4 6.20435 4.31607 5.44129 4.87868 4.87868C5.44129 4.31607 6.20435 4 7 4H17C17.7956 4 18.5587 4.31607 19.1213 4.87868C19.6839 5.44129 20 6.20435 20 7V17Z`;
 const warning = `M12 16a1 1 0 100 2 1 1 0 000-2zm10.67 1.47l-8.05-14a3 3 0 00-5.24 0l-8 14A3 3 0 003.94 22h16.12a3 3 0 002.61-4.53zm-1.73 2a1 1 0 01-.88.51H3.94a1 1 0 01-.88-.51 1 1 0 010-1l8-14a1 1 0 011.78 0l8.05 14a1 1 0 01.05 1.02v-.02zM12 8a1 1 0 00-1 1v4a1 1 0 002 0V9a1 1 0 00-1-1z`;
 
-const IconCompress = props => <Icon d={compress} {...props} />;
-const IconExpand = props => <Icon d={expand} {...props} />;
 const IconFilter = props => <Icon d={filter} {...props} />;
 const IconMoon = props => <Icon d={moon} {...props} />;
 const IconPause = props => <Icon d={pause} {...props} />;
@@ -27,11 +24,7 @@ const IconSignal = props => <Icon d={signal} {...props} />;
 const IconStop = props => <Icon d={stop} {...props} />;
 const IconWarning = props => <Icon d={warning} {...props} />;
 
-IconExpand.d = expand;
-IconCompress.d = compress;
-
 const iconNames = {
-  compress: `compress`,
   expand: `expand`,
   filter: `filter`,
   moon: `moon`,
@@ -46,7 +39,6 @@ const iconNames = {
 };
 
 const icons = {
-  compress: IconCompress,
   expand: IconExpand,
   filter: IconFilter,
   moon: IconMoon,
@@ -62,7 +54,6 @@ const icons = {
 
 export default icons;
 export {
-  IconCompress,
   IconExpand,
   IconFilter,
   IconMoon,
