+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { Card, Divider, Graph, Icon, Tag, VirtualList } from './atoms';
-export { Expand, IconsBar, JobEntry, JobGraph, Menu } from './molecules';
-export { AdminPanel, Jobs, Sidebar, TopActions } from './organisms';
+export { AdminPanel, Expand, IconsBar, JobEntry, JobGraph, Menu } from './molecules';
+export { Jobs, Sidebar, TopActions } from './organisms';
 export { Dashboard } from './templates/';
 
 // TODO: fix to indexes
