+++ src/components/molecules/AdminPanel/AdminPanel.stories.js
@@ -2,7 +2,7 @@ import { SB_LABELS } from '@constants';
 import AdminPanel from './AdminPanel.react';
 
 export default {
-  title: `${SB_LABELS.ORGANISMS}Admin Panel`,
+  title: `${SB_LABELS.MOLECULES}Admin Panel`,
 };
 
 export const Default = AdminPanel;
