+++ src/components/molecules/index.js
@@ -1,4 +1,4 @@
-export { default as Expand } from './Expand/Expand.react';
+export { default as AdminPanel } from './AdminPanel/AdminPanel.react';
 export { default as IconsBar } from './IconsBar/IconsBar.react';
 export { JobEntry, JobGraph } from './Job';
 export { default as Menu } from './Menu/Menu.react';
