+++ src/components/organisms/index.js
@@ -1,4 +1,3 @@
-export { default as AdminPanel } from './AdminPanel/AdminPanel.react';
 export { default as Jobs } from './Lists/Jobs/Jobs.react';
 export { default as Sidebar } from './Sidebar/Sidebar.react';
 export { default as TopActions } from './TopActions/TopActions.react';
